Add files via upload
here we can train images using the module named face_train.py and save the images and its labels in a yml file format using the recogniser and would store label ids using pickle into a file. In here all the faces are stored in a seperate folder and its read from the files using OS fn. 
In face recognition module we would take input from the webcam and try to recognise using LBPHFaceRecognizer , and also we will print the name on the face using puttext method . If the image is not recognised the name will not be shown in the window